feat: update README for medial axis extraction and enhance debug image output refactor: modify extract_ink_mask to return edges and adjust skeletonization process refactor: improve argument handling and code structure in args and main modules enhance: implement MedialAxis class for better skeleton representation fix: update debug counting to handle float values

+++ src/common/debug.py
@@ -4,7 +4,7 @@ from time import perf_counter
 
 _enabled = False
 _title = ""
-_steps: dict[str, int] = {}
+_steps: dict[str, float] = {}
 _times: dict[str, float] = {}
 _started = 0.0
 
@@ -24,7 +24,7 @@ def begin(title: str) -> None:
     _started = perf_counter()
 
 
-def count(step: str, amount: int = 1) -> None:
+def count(step: str, amount: float = 1) -> None:
     if not _enabled:
         return
     _steps[step] = _steps.get(step, 0) + amount
